Rename auth handler to login handler
